refactor: user-card type 파일 분리 #26
