+++ src/app/concert/[id]/_components/user-card/user-card.type.ts
@@ -1,0 +1,13 @@
+interface UserProps {
+  id: number;
+  name: string;
+  profileImage: string;
+  introduction: string;
+  transactionCount: number;
+}
+
+interface UserCardProps {
+  user: UserProps;
+  onClick: () => void;
+}
+export type { UserProps, UserCardProps };
